+++ web/app/components/style-guide/data/drop-down.jsx
@@ -1,0 +1,65 @@
+var React = require('react/addons');
+var nucleusReact = require('../../../../../assets/index');
+var DropDown = nucleusReact.components.DropDown;
+var Button = nucleusReact.components.Button;
+
+module.exports = {
+  name: 'Drop Down',
+  type: 'component',
+  overview: (
+    <p>
+      Drop down menu.
+    </p>
+  ),
+  properties: [],
+  examples: [{
+    description: (
+      <p>
+        Standard.
+      </p>
+    ),
+    example: (
+      <div>
+        just a test on a <DropDown
+          handle="simple"
+          content={
+            <ul className="plain-list">
+              <li>drop down item 1</li>
+              <li>drop down item 2</li>
+              <li>drop down item 3</li>
+            </ul>
+          }
+        /> drop down menu
+        <div>
+          <button>testing to make sure the hidden content does not effect other element below when hidden</button>
+        </div>
+      </div>
+    ),
+    exampleString: '<Badge>Standard</Badge>'
+  }, {
+    description: (
+      <p>
+        Right aligned.
+      </p>
+    ),
+    example: (
+      <div>
+        just a test on a <DropDown
+          align="right"
+          handle="simple"
+          content={
+            <ul className="plain-list">
+              <li>drop down item 1</li>
+              <li>drop down item 2</li>
+              <li>drop down item 3</li>
+            </ul>
+          }
+        /> drop down menu
+        <div>
+          <button>testing to make sure the hidden content does not effect other element below when hidden</button>
+        </div>
+      </div>
+    ),
+    exampleString: '<Badge>Standard</Badge>'
+  }]
+};